requirements.txt
streamlit>=1.28.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scipy>=1.10.0
bezier>=2023.7.28
